align: try to place inside viewport when there is no fallback
in case we don't have a fallback anymore (or never had one), try to
place the element inside the viewport by shifting not only along the
GrowDirection, but also on the perpendicular axis.

this makes e.g. dropdowns move left when there is not enough space on
the right side

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>