pwt-macros: builder: add `builder_cb` attribute
for generating the builder macros for callbacks.
When we want to generate a builder for a callback, we have to give
the paramter type explicitely, since we cannot really extract that
from the field type. (e.g. for the field Option<Callback<T>> we want
to use the IntoEventCallback<T> trait)

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>