widget: data_table: fix wobble for dropdowns
the first render pass renders the table with zero height, and adds
a scrollbar. This triggers a size change and we reposition the picker.

when the content is actually rendered, the scrollbar can vanish, and we
move the picker again (by the size of the scrollbar).

to avoid that, set the overflow to hidden when there is no table height
yet, since that omits the scrollbar and thus does not trigger a resize

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>