TabPanel: implement new TabBarStyle
Now the TabPanel has 3 styles:
* Current (default): Pills
* Material Primary
* Material Secondary

to achieve the animations of the material style tabs, we have to
manually position and size the indicator.

we also have to slightly change the internal structure of the
tabs to include a wrapper <span> so that we can find the size of the
text only

then we have to keep track of the indicator element, the whole button
and the text span so properly size and position the indicator.

while we're at it, clean up the code in the tabbar view function a bit.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>