dialog: auto center at first render
modal dialogs are centered by default, but via automatic css/browser
layouting. This is a disadvantage when the height of the dialog changes,
as it gets centered automatically again (resulting in a 'wobble').

With this, we use our alinging, which sets the position via a transform
that does not get touched by the browser whent he size changes
(so it grows to the bottom/end instead in all directions, resulting in
no 'wobble').

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>