input_panel: add keys to all elements
yews diff algorithm works best when all elements in a list have a key
assinged. Otherwise, when elements (dis)appear, some parts of the list
will be removed and added again, even if they have a key set.

To prevent this behaviour in the InputPanel, add a key to all elements
we add. For labels we reuse the label text as key (as that will most
likely not change), for fields we try to extract the key from it or
reuse the name. For custom fields we also try to extract the key.

If that did not work, auto-generate a key, but emit a warning, since
these auto-generated keys are most likely wrong. (They're only based
on the position, but when elements are added/removed that will change)

in addition, refactor the field adding since much of the code was
duplicated for the left/right/large fields

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>