align: change to using 'inset' instead of 'transform'
using 'transform' together with 'overflow:auto' makes problems with
nested use. instead of creating a new drawing context like
'position:fixed' promises, it will count towards the element and use
scrollbars for the overflowed content.

using inset isntead does not have this issue.
refactor the setting of that property in case we have to tweak this
behaviour again in the future.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>