align: revert to transform aligning
instead of setting the position directly. While that worked most of the
time. having variable size content inside would no longer work, as it
would change the text flow and layout (which is not what we want).

Using transform, we don't have that problem, but we have to make sure
that the aligned element does not have 'overflow: auto' but 'visible'
set, as nested elements will use the transformed element as the bounding
block (which will show scrollbars in that case).

This can happen for example when aligning dialogs with input fields that
have a tooltip that grows beyond the dialog itself.

To fix it for dialogs, we now nest the actual element below the <dialog>
element (which can have it's usual styling + width/height to 100%), but
the dialog itself has set 'overflow: visible'.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>