store: readd caching behaviour for update_filtered_data
commit 894d23c removed the version tracking, but that was integral to
caching the filtered/sorted data. Without that we always sorted/filtered
the list on read, even when scrolling with virtual scroll, etc.

this patch reintroduces the version tracking, but the only point where
we update the version is during the Drop handler of StoreWriteGuard, so
the new notify behaviour stays the same.

Fixes: 894d23c ("state/store: always notify, drop version tracking")
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>