input_panel: add custom label/field width properties
and add them to the style if set

note that the default input-panel width is set to 'min-content'. So when
one sets a custom width to something relative (e.g. 1fr) it will
probably be too small and one must also set 'width: auto' or similar
on the input-panel

also for now we have to maintain the default here and in the scss
(maybe there is a better way for that?)

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>