widget: data table: disable last enabled column
when deselecting columns, we don't want the user to deselect the last
one, since it's not possible to show a menu anymore to reselect any
column.

the only exception currently is when the column has the menu disabled.
we probably don't want to count such columns as 'active' for selection
purposes, but we currently don't have that information here.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>