widget: return `Row` from (error_)message instead of Html
with `Row` we are much more flexible, e.g. we can use all
builder style functions for classes/padding/margin/etc. and the only
downside is that sometimes we have to do an `.into()` to convert to Html

And even that is not always necessary since arguments of functions often
take something like: `Into<Html>` which `Row` implements.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>