props: add WidgetStyleBuilder trait
for having 'style' and 'set_style' functions on components that have
a WidgetBuilder already, makes it easier to set the style of
components, so instead of writing:

```
Component::new()
    .attribute("style", "style1: value1; style2: value2")
```

one can now do

```
Component::new()
    .style("style1", "value1")
    .style("style2", "value2")
```

making it much more obvious what is happening and easier if setting
multiple styles.

This also adds a shorthand for width and height, with a custom
CssLength enum that represents different css length units
and defaults to px if only a number is given

The CssStyles struct is there to hold the styles and generate the final
attribute string to use in components.

It implements a From trait for AsRef<str>, but that implmenetation is
not fully html/css spec conform as it leaves out some unused syntax.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>