widget: data table: fix keyboard focus regression
our recent change from `cell_class` to `cell_configuration` broke
keyboard navigation for tables, since the `tabindex` (among others)
attribute was missing on the cells.

To fix this, write a method to add attributes from other attributes, for
the WidgetStdProps, and use that for the cell_configuration instead of
overwriting the existing ones.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>