widget: form: add DisplayField
implements a widget that behaves like a field, but just for displaying
values. This way we can easily use that where we would normally use
fields.

This requires a value to set initially and has no way to change or
validate the value (that can all be done from the parent of this
component).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>