widget: data table: fix lost focus on scroll
when we selected a row/cell and scrolled away so that the element was
not in the dom, we lost focus and subsequent keypresses scrolled the
view, but did not select the next/previous element.

To fix that, we have to catch this case by checking if the element
behind the cursor still is in the visible range, and if not and there
was no new focus target, we refocus the scroll view.

This way we get the key press events again for the table.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>