widget: dropdown: rework dialog/focus handling
until now we used a modal dialog for the picker. While had some
advantages for focus/keyboard handling, it made it impossible to have
and 'editable' field while having the dialog open.

So instead, use a non-modal dialog and rework the focus/keyboard handling
to work as expected.

The biggest change here is how keyboard uses can navigate it:

instead of being 'captured' in the picker once it's open, using
tab/shift+tab simply closes the picker should the focus go outside the
current field

With this changes, having an editable dropdown works as expected, the
field gets the focus by default, the picker remains open until it's
closed or the widget loses focus, and one can open/close the picker
manually with the keyboard or trigger.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>