widget: number field: prevent invalid settings with scroll/keyboard
By don't in/decreasing the value if it's over the max/under the min
value. To insert a value when scrolling/keyboard even if none was set,
introducing a 'clamp_value' trait fn that simply forwards to the
underlying clamp with T::MIN and T::MAX as defaults. This is then used
to determine the initial value that is set.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>