widget: selection view: introduce VisibilityContext
this is used with a ContextProvider to provide a visibility information
to children elements, without the need to 'prop drill'.

Currently only used by a selection view, but could be used by all
elements that cache child elements that only show/hide them instead
of letting the dom change.

When using in a generic widget/etc, should react to higher level
visibility contexts.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>