widget: introduce FocusTracker
this is a helper struct to better handle focusin/out events.
We had the pattern with DelayedFocusChange in a few components, and they
were basically all doing the same.

To factor out, we have to use a bit of a different approach (see the
comment on the struct itself), since we don't want to introduce an extra
component in the component tree for this (since nothing is really
rendered for that).

It's not saving us much code for now, but with every new component
needing that, we can make sure we have the same behavior.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>