dropdown: fix picker not closing on first select after input
we don't want to close the picker when we change the value via keyboard
input, so we save that info in `change_from_input`.

Since we only handle that flag when the picker is open, we must only set
it when the picker is open too. Otherwise we possibly set it while the
picker is closed, having the wrong behavior when the picker gets opened
afterwards.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>