widget: boolean: improve html markup to be more sensible
Currently we "misused" the tooltip as part of the markup, but put a
container around for css purposes.

Instead make the actual checkbox/switch markup completely independent
from the tooltip and only move the onclick handler there.

This way it's more easy to add other things like a 'boxLabel' later.

While we're changing the markup anyway, remove the use of the html!
macro and use our builtin builder classes for containers and icons.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>