widget: move FieldLabel functionality into Boolean
We only ever need labels on the right side (called box label now), only
on checkboxes, so it does not really make sense to have a separate
widget for that.

For now we only accept text as the label, but we could change that
later if we need that.

This makes the 'FieldLabel' available and ready to use elsewhere, e.g.
the input panel.

For convenience, we implement some From traits for it, so one can use a
bare string to create such a field label.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>