widget: form: add RadioButton component
instead of reusing the Checkbox component for creating radio buttons,
have a separate component for that. This makes it easier to specialize
the code for that use case.

The code is mostly a mix of Checkbox and Boolean (for some of the newer
features), with the irrelevant portions stripped out.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>