widget: data table: react to column changes
the headers are part of the properties, so if they change, we should
react to that. Until now, we only accepted them at the first 'create'
call, which meant that when the columns changed afterwards (e.g. when we
want to pass different data to the renderer) it did not get rerendered
with the new information.

To do this, refactor the header processing code from 'create' into it's
own function, and call it also when the headers changed in the 'changed'
method.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>