widget: dropdown: use popover instead of a dialog
by using popovers, we don't have get the messy focus behaviour of
dialogs, for example those focus the first element when the dialog is
opened, which often is the header menu of a datatable.

This also makes the code a bit shorter, since we can restore the focus
before the popover is fully closed (and simply hide it as soon as
possible).

for now we have to use our own bindings to show/hide/togglePopover, but
this can be removed with a later web-sys version.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>