state: loader: don't unnecessarily notify listeners
in the `load` method, we just setup the loading itself, but the write
guard does that on every `Drop`. To prevent that and only trigger
the `on_change` callback when the data really changed, combine the
calls to `write()` here so we only have one guard and disable the
notification on it.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>