theme density: rework naming to more telling variants
Rework the naming and doc comments of the theme density type but
include backward compatibility for the old variants when parsing the
density from a string. There's nothing to adapt on the actual UI
examples or products.

The bigger improvement is avoiding "Auto" for the default theme
spacing/density, as "auto" is widely used for something that
dynamically adapts on some environment aspects. But all our themes do
not have any such queries for high/low dpi or screen width to choose a
fitting spacing dynamically. Use `Preset` to avoid clashing with the
widely used `Default` trait from the rust std lib, albeit we could use
that if we really want.

The smaller fine tuning is renaming "High" to "Compact" and "Touch" to
"Relaxed".
The reasoning for the first is that most people, including our user
base, does not work with density a lot in their every day life, be it
physical density or screen density, so "higher density", while
definitively correct for what it does, is not that intuitive, i.e.
when "higher" makes things "smaller" we have a sort of double-negation
that is almost never good to have for UX.
The reasoning for the second (Touch -> Relaxed) is that this is not
doing anything with touch directly, it just might correlate with that
use case, and coupling a setting name with one specific use case is a
bit confusing for peoples coming from other use cases, like simply
having a higher DPI but not doing desktop wide scaling.

Note the naming was initially suggested by me with Thunderbird as
example, so this is really just improving on my own short comings from
the past. It also seems that newer Thunderbird Versions (e.g., latest
134.0b4) changed the wording to something similar than we now have
after this commit.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>