widget: display: don't set inital value
when doing that, a new value that is set via `new` won't get used as a
value anymore, which is undesired. only use the value for the default,
which was the original intenion.

This fixes an issue where the value of a display field changes after the
first creation and would not update it's displayed value.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>