widget: dropdown: update picker placement after showing
It seems that under certain circumstances (e.g. a dropdown in a
dialog) updating the position of the picker before it's opened (it's a
popover) does not work properly in safari/webkit/epiphany.

When trying to open such a dropdown, it (or some parts of it) would be
rendered with a height of 0px, even though the elements have a height
set explicitly.

To fix this, update the picker placement after opening it. What
exactly is the cause (e.g. if it's a browser bug or our css/aligning)
is yet to be determined.

This fixes an issue on webkit (tested with epiphany) that opening
dropdowns in a dialog did not show the dropdown properly.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
 [ TL: extend comment to mention webkit explicitly ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>