selection/store/tree_store: use `where` for generics in hooks
instead of in-lining them in the function definition. the `hook!` macro
will add a `'hook` liftime via a `where` clause to them when expanded,
which triggers the clippy lint `multiple_bound_locations` [1] as the
generics would then be bound in the function signature and the `where`
clause. so just define them in the `where` clause right away to avoid
that.

[1]:
https://rust-lang.github.io/rust-clippy/master/index.html#multiple_bound_locations