macros: fix build for tests
it seems adapting the test case in pwt-macros was forgotten at some
point. this makes the test cases build-able again. please note that
this does not make all test cases here work, as some of them seem to
expect the output of specific rustc versions. this should probably be
addressed in a separate commit.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>