state: selection: make selection interface more consistent
due to the interior mutability of th Selection (it uses Rc<RefCell<>>),
we don't need the selection to be mutable when we select something. This
is reflected in in the interface for e.g. `select()` and `toggle()`
which only take a `&self`. For `bulk_select()` the same logic can be
applied, but it still took a `&mut self` which was not a problem, but it
was unnecessary. To make it more consistent, simply change from `&mut
self` to `&self`.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>