dialog: put all dialogs into a portal
otherwise, the dialog is part of the markup where we insert it, which
can make problems with event handling, for example:

Creating a form that can open a dialog which contains also a form,
produces two problems:

* nesting forms is explicitly not allowed [0]
* if the form has a 'submit' input, pressing 'enter' in either form
  will submit the upper one.

To fix such issues in general, put the dialog into portal, which
put it directly into the body, where it can't interfere with other
elements up the tree. The yew component tree is not affected.

0: https://html.spec.whatwg.org/multipage/forms.html#the-form-element

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>