widget: data table: prevent duplicate yew keys for rows
Yews diff algorithm assumes that elements in a list have unique keys, so
it can properly diff the new and old list, but our Store/Datastore does
not guarantee that records have unique keys.

For the data table rows, we used the record keys as yew keys, so this is
a problem when keys in the store are duplicated.

To prevent an application wide panic, modify keys for DataTableRows that
are to be displayed and have duplicated keys.

In debug mode, log a prominent warning, so it's easier to detect.

This is more in depth prevention of crashes/problem like in bug #6382.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>