touch: gesture detector: implement a touch only mode
when dealing with touch devices, using pointer events is not practical,
since those will be canceled by touch events. While doing

```
event.prevent_default();
```

inside the 'touchstart' event would allow us to use pointer events,
click events on touch enabled devices would not anymore (since those are
generated from the touch events when touching).

As a slightly less broken workaround, detect if the browser is touch
capable, and set the gesture detector to a touch only mode, so that it
only uses the touchstart/end/move/cancel events.

This makes it both work on a touch enabled device with touch, and a
non-touch enabled device with the mouse. One downside is that it does
not work with a mouse on touch enabled devices, but this should not that
big of a problem, since the gesture detector is intended to be used with
touch interfaces in the first place.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>