widget: form: number: set 'inputmode' attribute
Even for Number fields, the 'type' attribute is set to 'text' to handle
invalid value/state properly. To help mobile browsers use the correct
software keyboard (e.g. a special one for number input) use the
'inputmode' attribute[0].

For this we have to add an 'is_decimal' function to the NumberTypeInfo
trait, so we can decide if we want to set 'decimal' or 'numeric'.

0: https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/input#inputmode

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>