state: keyed slab tree: add key -> node_id cache
The methods `find_node_by_key` and `find_node_by_key_mut` have to
linearly search for the key in the tree. When doing that for a large
number of keys, this becomes very slow.

To speed that up, introduce a `key_cache`, which is just a
`HashMap<Key, uszie>` that points to the node_id in the slab tree.

We have to be careful to update this cache with every insert or remove
operation.

* on insert, add the new entry into the cache
* on removal, remove the entry and do it recursively for all children of
  that node
* on bulk append: iterate over the newly added subtree and add the keys
* on tree/root replace: recalculate the cache for the whole tree.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>