catalog loader: defer first content rendering after language is loaded
Currently, whenever the catalog loader tries to load a language, it
will empty out the content once, iow. render an empty body and then the
actual content again. This means that all sub components are recreated
and lose their state.

When changing between languages this is necessary, since otherwise
already rendered text will not be rerendered with the new language.

On first page load on the other hand, rendering once in the default
language and then immediately in the target language (even if it's also
the default language) does not make much sense, since
* most of the time the language catalog is rather small
  (e.g. pdm german mo file is ~75KiB), so the initial rendering is immediately
  lost and rerendered
* even if the language would take longer to load, displaying it for a
  relatively short time in the wrong language is not helpful.
* doing anything in a 'create' method of components that should only run
  once, cannot be done in a way that retains some info about it. For
  example the PMG mobile gui that tries to do some action from the GET
  parameters, but only once (including removing the GET parameters).
  That can work, but we can't show any popup/snackbar then since the
  language loaded in the meantime and we have no knowledge that we did
  such an action.

To fix this, render nothing until any language was at least loaded once.

On our older ExtJS GUIs, the server determines the language with a
cookie and the index has the link in the <head> section of the html.
This will block the javascript execution until that loaded, so this
solution here mimics the behavior on our ExtJS GUIs.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>