widget: catalog loader: fix behavior when no language is set
Recently, a change in the catalog loader was introduced to defer loading
of the page until after the language was loaded. This worked, except in
the case that no language was set at all, neither via a property nor via
the browser local storage.

When that occurred, there was no attempt to load the catalog but also
the indicator that we're finished with loading was not set, so the child
elements were never displayed.

To fix this, treat an empty language string the same as if it were the
'default_lang', which then triggers the initialization, etc. correctly.

Fixes:  3a437d1 (catalog loader: defer first content rendering after language is loaded)
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>