store: add helper methods to extract a record via a selection
this adds helpers to the `Store` (and its `StoreState`) that have
similar objectives:

- `selected_record(Selection) -> Option<T>`: returns a clone of the
record that has the same `Key` as the currently selected item in the
`Selection`. this requires the type `T` to implement `Clone`.
- `lookup_with_selection(Selection) -> Option<&T>`: requires the user
to acquire a read guard on the store first and will then return a
reference to the record of the store that matches the currently
selected item in the `Selection`.
- each also has variation that is prefixed with `multi_` and returns a
vector of all entries that have a key selected by a `Selection` in
multi-select mode

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251121161423.390869-2-s.sterz@proxmox.com