pwt-macros: builder: allow multiple generic types in callback variant
if there is a callback type with multiple generic variants, e.g.
Foo<A,B> we currently can't use it since we only expect one type
in the macro.

Extend the parser to add multiple types for the generic part.

So we can now write for example:

 #[builder_cb(IntoEventCallback, into_event_callback, A, B)]
 foo: Callback<A, B>,

This way it's compatible with the current way we write it.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>