props: make RenderFn's return type generic
and use that with the type `String` instead of our TextRenderFn, that is
now obsolete.

This not only gets rid of a trait, but also allows us to have more
generic render functions in use, e.g. if some component needs a
render function that returns e.g. a `Dialog` to do something with it,
which would now be possible.

By defining the generic type `OUT` as `Html` by default, we don't break
compatibility with existing users.

This is now very similar to yew's `Callback` type, but it just gives a
reference to the function instead of an owned type.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>