widget: data table: don't fire extra selection change event
Instead of always clearing the selection and toggling the correct one,
directly select the correct one if the selection is not in multiselect
mode.

This prevents multiple selection change events when a user clicks a line
in the data table. (namely one from the clear, and one from the toggle)

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>