plain date: improve format parser
by using more idiomatic rust code, and also fixes some bugs that are now
tested with additional test cases.

In particular this changes the parser from being purely index based to
using a mutable slice for proceeding, and an iterator + match for
handling the different format characters.

The new tests are for doing early error checking e.g. in case the format
string contains multiple variants of the same type, and requires valid
values for those.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>